Set default change result to empty list
The restarting script will be skipped if it's not a debian system,
however if we skipped it at the point of `include` in 1.9.x, the
rest of when conditions will not be evaluated, however the
`with_item` will still try to eval the template, this will cause
error when the result actually not exists.
In this commit, we set all default to empty list, so it won't be
an error anymore.